Added and coded menu methods and constructor methods

+++ src/FullTimeTeacher.java
@@ -2,7 +2,7 @@ public class FullTimeTeacher extends Teacher {
     private int experienceYears = 0;
 
     // Constructor
-    public FullTimeTeacher(String name, int id, int baseSalaryPerHour, int experienceYears) {
+    public FullTimeTeacher(String name, int id, double baseSalaryPerHour, int experienceYears) {
         this.name = name;
         this.id = id;
         this.baseSalaryPerHour = baseSalaryPerHour;
